move k6Client
